+++ objectbox/CHANGELOG.md
@@ -48,6 +48,8 @@
   ```
   See the `Store` documentation for details.
 * Add `Store.removeDbFiles()` to conveniently delete database files or an in-memory database.
+* Add `Store.dbFileSize()` to get the size in bytes of the main database file or memory occupied by
+  an in-memory database.
 
 ## 2.4.0 (2023-12-13)